pyproject.toml
[tool.poetry]
name = "dnscatz"
version = "0.2.1"
description = "DNS Catalog Zones Tools"
authors = ["Jakob Schlyter <jakob@kirei.se>"]
license = "BSD-2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
dnspython = "^2.2.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
pytest = "^8.2.1"
pytest-isort = "^3.0.0"
ruff = "^0.4.6"
pytest-ruff = "^0.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
catz2nsd = "dnscatz.catz2nsd:main"
zones2catz = "dnscatz.zones2catz:main"

[tool.isort]
profile = "black"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = ["E501", "I001", "SIM102"]
